requirements.txt
fastapi==0.116.0
uvicorn==0.35.0

# JWT handling
PyJWT==2.8.0

# Password hashing
passlib[bcrypt]==1.7.4

# Pydantic and related
pydantic==2.11.7
pydantic_core==2.33.2
annotated-types==0.7.0
typing-inspection==0.4.1
typing_extensions==4.14.1

# FastAPI dependencies
starlette==0.46.2
anyio==4.9.0
sniffio==1.3.1
h11==0.16.0
httptools==0.6.4
websockets==15.0.1
uvloop==0.21.0
watchfiles==1.1.0
click==8.2.1

python-dotenv==1.1.1
PyYAML==6.0.2
idna==3.10
itsdangerous
authlib
httpx
redis
databases
sqlalchemy
asyncpg
psycopg2-binary